requirements.txt
langchain
langchain-community
langchainhub
sqlalchemy
pydantic
streamlit
nest-asyncio
tiktoken
python-dotenv
langchain-google-genai
google-generativeai
chromadb